requirements/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    ./compile-requirements.sh
#
-e git+https://github.com/jaywink/django-markdownx.git@bd2fa9dffcbfccc62f9cef2677921537f100bbaa#egg=django-markdownx==2.0.21.1
-e git+https://git.feneas.org/jaywink/federation.git@979b43636d61cb667af78549178958848dcd0241#egg=federation==0.18.0-dev.1
-e git+https://github.com/jaywink/pyembed.git@6f8c1cc98d61ee3083e9803255e4b2cc90b5a0dd#egg=pyembed==1.3.3.1
arrow==0.12.1
asgi-redis==1.4.3
asgiref==1.1.2            # via asgi-redis, channels, daphne
attrs==18.1.0             # via automat, twisted
autobahn==18.7.1          # via daphne
automat==0.7.0            # via twisted
backcall==0.1.0           # via ipython
backports-abc==0.5
beautifulsoup4==4.6.1
bleach==2.1.3
certifi==2018.4.16        # via requests
channels==1.1.8
chardet==3.0.4            # via requests
circus==0.15.0
click==6.7                # via rq
commonmark==0.7.5
constantly==15.1.0        # via twisted
coreapi==2.3.3            # via django-rest-swagger, openapi-codec
coreschema==0.0.4         # via coreapi
croniter==0.3.25          # via rq-scheduler
cssselect==1.0.3
daphne==1.4.2             # via channels
decorator==4.3.0          # via ipython, traitlets
defusedxml==0.5.0         # via python3-openid
dirty-validators==0.4.0
django-allauth==0.36.0
django-autoslug==1.9.3
django-braces==1.13.0
django-crispy-forms==1.7.2
django-dynamic-preferences==1.5.1
django-enumfields==0.10.1
django-environ==0.4.5
django-extensions==2.1.0
django-floppyforms==1.7.0
django-fsm==2.6.0
django-haystack==2.8.1
django-js-reverse==0.8.2
django-memoize==2.1.1
django-model-utils==3.1.2
django-redis==4.9.0
django-rest-swagger==2.2.0
django-reversion==3.0.0
django-robots==3.1.0
django-rq==1.2.0
django-secure==1.0.1
django-settings-export==1.2.1
django-tables2==1.21.2
django-versatileimagefield==1.9
django==2.0.8
djangorestframework==3.8.2
future==0.16.0            # via commonmark
html5lib==1.0.1           # via bleach
hyperlink==18.0.0         # via twisted
idna==2.7                 # via hyperlink, requests
incremental==17.5.0       # via twisted
ipdata==2.6
ipython-genutils==0.2.0   # via traitlets
ipython==6.5.0
isodate==0.6.0            # via python-xrd
itypes==1.1.0             # via coreapi
jedi==0.12.1              # via ipython
jinja2==2.10              # via coreschema
jsonschema==2.6.0
lxml==4.2.2
markdown==2.6.11
markupsafe==1.0           # via jinja2
msgpack-python==0.5.6     # via asgi-redis
oauthlib==2.1.0           # via requests-oauthlib
openapi-codec==1.3.2      # via django-rest-swagger
parso==0.3.1              # via jedi
persisting-theory==0.2.1  # via django-dynamic-preferences
pexpect==4.6.0            # via ipython
pickleshare==0.7.4        # via ipython
pillow==5.0.0
prompt-toolkit==1.0.15    # via ipython
psutil==5.4.6             # via circus
psycopg2-binary==2.7.5
ptyprocess==0.6.0         # via pexpect
pycryptodome==3.6.6
pygments==2.2.0           # via ipython
pyhamcrest==1.9.0         # via twisted
python-dateutil==2.7.3    # via arrow, croniter
python-opengraph-jaywink==0.2.0
python-xrd==0.1
python3-openid==3.1.0     # via django-allauth
pytz==2018.5
pyzmq==16.0.4             # via circus
raven==6.9.0
redis==2.10.6
requests-oauthlib==1.0.0  # via django-allauth
requests==2.19.1          # via coreapi, django-allauth, python-opengraph-jaywink, requests-oauthlib
rq-scheduler==0.8.3
rq==0.12.0
simplegeneric==0.8.1      # via ipython
simplejson==3.16.0        # via django-rest-swagger
six==1.11.0               # via asgi-redis, asgiref, autobahn, automat, bleach, circus, django-dynamic-preferences, django-extensions, html5lib, isodate, prompt-toolkit, pyhamcrest, python-dateutil, traitlets, txaio, unicode-slugify
tornado==4.5.3            # via circus
traitlets==4.3.2          # via ipython
twisted==18.7.0           # via daphne
txaio==18.7.1             # via autobahn
typing==3.6.4
unicode-slugify==0.1.3
unidecode==1.0.22         # via unicode-slugify
uritemplate==3.0.0        # via coreapi
urllib3==1.23             # via requests
uwsgi==2.0.17.1
wcwidth==0.1.7            # via prompt-toolkit
webencodings==0.5.1       # via html5lib
whitenoise==4.0
whoosh==2.7.4
zope.interface==4.5.0     # via twisted
